container restructure
