+++ frontend/src/Containers/Footer/styles.js
@@ -4,7 +4,7 @@ export const FooterWrapper = styled.section`
   width: 100%;
   display: flex;
   flex-flow: column;
-  padding: 0 0 1em 0;
+  padding: 0 0 1rem 0;
   overflow-wrap: break-word;
   text-align: center;
 `
