+++ frontend/src/Containers/Header/styles.js
@@ -1,7 +1,6 @@
 import styled from "styled-components"
 
 export const HeaderWrapper = styled.section`
-  width: 100%;
   display: flex;
   flex-flow: column;
   overflow-wrap: break-word;
